rename status_bar files
