+++ CMakeLists.txt
@@ -58,7 +58,7 @@ set(ADTOOL_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/contents_widget.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/main_window.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/entry_widget.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/src/status_bar.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/status.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/members_model.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/members_widget.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/src/entry_model.cpp
